A new controller class named `CompaniesController` has been created in the `CompanyEmployees.Presentation.Controllers` namespace. This class inherits from the `ControllerBase` class provided by the `Microsoft.AspNetCore.Mvc` namespace.
The `CompaniesController` class has been decorated with the `[ApiController]` attribute. This attribute indicates that the class is a controller that uses the API Controller conventions of ASP.NET Core.

The `[Route("api/companies")]` attribute has been added to the `CompaniesController` class. This attribute sets the route template to which this controller responds. The controller will handle HTTP requests that are sent to the `api/companies` URL path.

The `using Microsoft.AspNetCore.Mvc;` directive has been added at the top of the file. This directive allows the code to use the `ControllerBase` class and the `ApiController` and `Route` attributes without needing to fully qualify their names.